Update: Renamed modules directory to pages
