+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React, { ReactElement } from 'react'
 import { Route, BrowserRouter as Router, Switch } from 'react-router-dom'
 import AppContext from './AppContext'
-import Blog from './modules/Blog'
-import Home from './modules/Home'
+import Blog from './pages/Blog'
+import Home from './pages/Home'
 import data from './data'
 import './common/styles/app.css'
