--- src/modules/Home/Sections/index.ts
@@ -1,4 +1,0 @@
-export { default as Footer } from './Footer'
-export { default as Projects } from './Projects'
-export { default as Testimonials } from './Testimonials'
-export { default as WelcomeScreen } from './WelcomeScreen'
