+++ src/pages/Home/Projects/Modal/ModalMobile/ModalMobile.tsx
@@ -2,7 +2,7 @@ import { IoIosArrowBack, IoIosArrowForward } from 'react-icons/io'
 import React, { FunctionComponent, ReactElement, useState } from 'react'
 import { BsBuilding } from 'react-icons/bs'
 import { HiUserGroup } from 'react-icons/hi'
-import { IProject } from '../../../../../../common/interfaces'
+import { IProject } from '../../../../../common/interfaces'
 import { MdClose } from 'react-icons/md'
 import ModalTeam from '../ModalTeam'
 import ProjectInfo from '../../ProjectInfo'
