+++ src/pages/Home/Projects/Modal/index.ts
