+++ src/pages/Home/Projects/ProjectInfo/ProjectInfo.tsx
