+++ src/pages/Home/Projects/ProjectInfo/index.ts
