+++ src/pages/Home/Projects/ProjectInfo/styles.css
