+++ src/pages/Home/Projects/ScrollAnimation/index.ts
