+++ src/pages/Home/WelcomeScreen/WelcomeScreen.tsx
@@ -1,6 +1,6 @@
 import React, { EffectCallback, FunctionComponent, ReactElement, useEffect } from 'react'
 import ViewMyWorkButton from './ViewMyWorkButton'
-import { deejay250x250 } from '../../../../common/images'
+import { deejay250x250 } from '../../../common/images'
 import { executeTitleTypeAnimation } from './helpers'
 import './styles.css'
