+++ src/pages/Home/WelcomeScreen/helpers.ts
